+++ R/areainfo.R
@@ -7,6 +7,7 @@
 #' @param theta_lim range of integration to be computed
 #' @param which.items an integer vector indicating which items to include in the expected information function.
 #'   Default uses all possible items
+#' @param ... additional arguments passed to \code{\link{integrate}}
 #'
 #' @return a \code{data.frame} with the lower and upper integration range, the information area
 #'   within the range (Info), the information area over the range -10 to 10 (Total.Info), proportion
@@ -39,8 +40,8 @@
 #'
 #' }
 areainfo <- function(x, theta_lim, which.items = NULL, ...){
-    f <- function(theta, mod, which.items)
-        testinfo(x=mod, Theta=matrix(theta), which.items=which.items)
+    f <- function(theta, x, which.items)
+        testinfo(x=x, Theta=matrix(theta), which.items=which.items)
     if(missing(x)) missingMsg('x')
     if(missing(theta_lim)) missingMsg('theta_lim')
     stopifnot(length(theta_lim) == 2L)
@@ -48,8 +49,8 @@ areainfo <- function(x, theta_lim, which.items = NULL, ...){
 
     nitems <- ifelse(is.null(which.items), extract.mirt(x, 'nitems'), length(which.items))
     ii <- integrate(f, lower = theta_lim[1L], upper = theta_lim[2L],
-                    mod=mod, which.items=which.items, ...)
-    iT <- integrate(f, lower = -10, upper = 10, mod=mod, which.items=which.items, ...)
+                    x=x, which.items=which.items, ...)
+    iT <- integrate(f, lower = -10, upper = 10, x=x, which.items=which.items, ...)
     ret <- data.frame(LowerBound=min(theta_lim), UpperBound=max(theta_lim),
                       Info=ii$value, TotalInfo=iT$value, Proportion=ii$value/iT$value,
                       nitems=nitems)